+++ lib/vcr/structs.rb
@@ -1,7 +1,39 @@
 require 'forwardable'
 
 module VCR
+  module BodyNormalizer
+    def initialize(*args)
+      super
+      normalize_body
+    end
+
+    private
+
+    def normalize_body
+      # Ensure that the body is a raw string, in case the string instance
+      # has been subclassed or extended with additional instance variables
+      # or attributes, so that it is serialized to YAML as a raw string.
+      # This is needed for rest-client.  See this ticket for more info:
+      # http://github.com/myronmarston/vcr/issues/4
+      self.body = case body
+        when nil, ''; nil
+        else String.new(body)
+      end
+    end
+  end
+
   module HeaderNormalizer
+    # These headers get added by the various HTTP clients automatically,
+    # and we don't care about them.  We store the headers for the purposes
+    # of request matching, and we only care to match on headers users
+    # explicitly set.
+    HEADERS_TO_SKIP = {
+      'connection' => %w[ close Keep-Alive ],
+      'accept'     => %w[ */* ],
+      'expect'     => [''],
+      'user-agent' => ["Typhoeus - http://github.com/pauldix/typhoeus/tree/master", 'Ruby']
+    }
+
     def initialize(*args)
       super
       normalize_headers
@@ -9,18 +41,30 @@ module VCR
 
     private
 
+    def important_header_values(k, values)
+      skip_values = HEADERS_TO_SKIP[k] || []
+      values - skip_values
+    end
+
     def normalize_headers
       new_headers = {}
 
       headers.each do |k, v|
-        new_headers[k.downcase] = case v
+        k = k.downcase
+
+        val_array = case v
           when Array then v
           when nil then []
           else [v]
         end
+
+        important_vals = important_header_values(k, val_array)
+        next unless important_vals.size > 0
+
+        new_headers[k] = important_vals
       end if headers
 
-      self.headers = new_headers
+      self.headers = new_headers.empty? ? nil : new_headers
     end
   end
 
@@ -55,9 +99,24 @@ module VCR
     end
   end
 
+  module StatusMessageNormalizer
+    def initialize(*args)
+      super
+      normalize_status_message
+    end
+
+    private
+
+    def normalize_status_message
+      self.message = message.strip if message
+      self.message = nil if message == ''
+    end
+  end
+
   class Request < Struct.new(:method, :uri, :body, :headers)
     include HeaderNormalizer
     include URINormalizer
+    include BodyNormalizer
 
     def self.from_net_http_request(net_http, request)
       new(
@@ -74,6 +133,8 @@ module VCR
   end
 
   class ResponseStatus < Struct.new(:code, :message)
+    include StatusMessageNormalizer
+
     def self.from_net_http_response(response)
       new(response.code.to_i, response.message)
     end
@@ -81,17 +142,7 @@ module VCR
 
   class Response < Struct.new(:status, :headers, :body, :http_version)
     include HeaderNormalizer
-
-    def initialize(*args)
-      super
-
-      # Ensure that the body is a raw string, in case the string instance
-      # has been subclassed or extended with additional instance variables
-      # or attributes, so that it is serialized to YAML as a raw string.
-      # This is needed for rest-client.  See this ticket for more info:
-      # http://github.com/myronmarston/vcr/issues/4
-      self.body = String.new(body) if body
-    end
+    include BodyNormalizer
 
     def self.from_net_http_response(response)
       new(